EN + CZ -- unused methods removed:
i.e. the ones reported in translator_report.txt as...

======================================================================
WARNING: The following translator methods are declared in the
Translator class but their identifiers do not appear in source files.
The situation should be checked. The .cpp files and .h files excluding
the '*translator*' files in doxygen/src directory were simply searched
for occurrence of the method identifiers:

  QCString idLanguageCharset()
  QCString trAlphabeticalList()
  QCString trDCOPMethods()
  QCString trDirDependency(const char *)
  QCString trFuncProtos()
  QCString trFunctionPrototypeDocumentation()
  QCString trSearchForIndex()
  QCString trWrittenBy()