Bug 706520 - Fortran: in body documentation lands on wrong place
In case we are in the body of a subroutine or function set the in body flag.
When comment has been handled (each Fortran comment block is handled as one block, not as parts) reste the in body flag.
In case we have to handle documentation of an argument, reset the inbodyDocs field to the empty string. The documentation stored was no inbody documentation but argument documentation).