Debug output for lexical analyzer
The previous commit (pull request) was missing the debug output possibility for config.l and partial for doctokinizer.l.
The later is called on different places by means of the generic definition doctokinizerYYlex, which is declared by means of YY_DECL.
To be able to add the debug output the YY_DECL had to be redefined to be a new function and the new function, with the debug output possibility, has to be called through the original function name.