dot: use threads for file patchers
Added a new base class, DotTask, that DotRunner and DotFilePatcher
inherits. This enable reuse of the existing threading code.

Also, when adding a new file patcher, the .svg are pushed in front
of the list, so that files dependant on svg include the latest version
of it.

QDir::current().rename() was replaced with Qdir.rename, since
QDir::current is not thread safe

Signed-off-by: Adrian Negreanu <adrian.m.negreanu@intel.com>