Bug 721302 - [Latex/PDF] Merging brief and detailed description in file section
Not only for Man pages the extra returns are necessary but also for LaTeX output.
The change in filedef.cpp is a direct consequence of t the bug report.
The groupdef.cpp has been tested on propriety code and the namespacedef.cpp and dirdef.cpp directly follow from code similarity.