Incorrect handling dependencies
The handling of a number of dependencies was incorrect, it was stated that the i(source) cpp file depended on the (included) h file, this should be that the resulting (object) o file depends on the cpp and the h file. See e.g. the layout.cpp dependency
Some dependencies used ../generated_src/doxygen it is better to use the definition GENERATED_SRC