Add line numbers in case comment is not closed properly.
This bug fix is an extension to the bug fix:  Add warning when encountering a nested comment start (/*) without matching end (*/).
It is quite hard to find the place in the source file where the problem occurs. This bug fix helps to locate part of the problems, it is not perfect it only gives the starting comment tags that give a problem (other problems also occur e.g. when // is followed by */).