Limit number of dot threads
In case doxygen is run with DOT_NUM_THREADS = 0 the number of available processors + 1 is used for the number of threads. In case the number of threads is set to a value this value is used irrespective of the number of available processors.
With this patch the number of threads is limited to the number of available processors +1.