Showing error message on windows in case on error on startup
On case, on windows, more than 1 argument is given doxywizard stops without showing a message even though a printf is present.
In this patch the message is show (for all platforms) by means of a message box, furthermore in case the argument --help is given the usage message is given in a message box.