Extension in config.xml if type=string with format=image
In case of a project logo a preview is shown in the "wizard mode". In "expert mode" only the name is shown.
In this patch the config.xml the type=string possibilities are extended with format=image (config.xml, configgen.py, config.h).
The doxywizard has been extended so that in "expert mode" this type is supported and that a preview is given (expert.cpp, inputstring.cpp and inputstring.h), furthermore in wizard.cpp a minimum label size has been defined to prevent jumping of the items in case of switching between a message text and an icon.