Coded coloring of flow statements corrected (regression)
Which patch 8e8c9a1 (dd. May 6) the rule:
<Start>{FLOW}/[,( \t\n].*
has been changed into"
<Start>{FLOW}/[,( \t\n]
(analogous: <Start>{COMMANDS}/[,( \t\n])
To overcome the warning: fortrancode.l:744: warning, dangerous trailing context.

The result is that the code coloring of (e.g.) the if statements in:
      subroutine test_on_if
        if (a == b) then
          c = d
        endif
        if(a == b) then
          c = d
        endif
      end subroutine
was removed, due to the fact that the rule: <Start>{ID}{BS}/"(" was used instead of the flow coloring rule.

This patch fixes the rules so the lines are colored again and there is no warning.