Bug 700510 - Bibliography duplicate filenames fails
Problem in the bug report is that the bib files in the different directories have the same name and thus overwrite each other.
This patch solves this problem as well as the possibility to have spaces in the file name.
For all output formats the bib files are copied to a new numbered name: "bibTmpFile_<number>", for Latex on the LATEX_OUTPUT directory and for all other formats in the subdirectory "bibTmpDir" of the OUTPUT_DIRECTORY.
The files for Latex are retained whilst the files for the other formats are removed after usage (as it was before).