Messages truncated in warnings file
The current limit for a message length in the messages file is 4095 characters. Most of the time this is more than sufficient, but in case of Fortran and a lot of arguments this might not be sufficient.
With Fortran the "data type" can be quite long as it can also includes words like: double precision, dimension, intent(inout) etc. When we have a lot of arguments and just one is described it also tries to write out the names of the non-described parameter/ arguments with the word parameter prepended resulting in an even longer line.

This patch increases the size of the buffer.