Support multiple extra HTML stylesheets.
In regular use it happens that it is more convenient to have multiple extra HTML stylesheets, e.g. in case of different levels of requirements: company, department, project
With this patch it is possible to specify multiple HTML stylesheets.
Note: for compatibility the configuration parameter HTML_EXTRA_STYLESHEET has not been renamed to HTM_EXTRA_STYLESHEETS.