Bug 735745 - Spurious ASSERT message
The ASSERT message is replaced with a normal message.
The cmdName might be overwritten so we have to save the name cmdName first, also the first test should not try to read any further, but jump to the end as well )a push has already been performed so a pop etc. is necessary)