Different latex problems
- Bibliography refers to wrong page in the toc
In case of LaTeX the reference in the toc tab to the Bibliography page is not correct.
According to e.g. http://lookherefirst.wordpress.com/2008/10/05/add-the-bibliography-in-the-table-of-contents/
when using the hyperref package the reference of the toc should be after the \bibliography entry.
See to it that the Bibliography page starts at an odd page (like all other chapters)

- See to it that the Index page starts at an odd page (like all other chapters)

- Not all items are shown in the Index / Index generation crashes (Bug_733323) / Index is not shown
The \hypertarget definition contains on 2 places the entire string (including section references and index references) in the caption part.
In all other places the caption is empty and the caption part is not used (as a reference) in the \hyperlink command. The caption has been made empty/