Empty entry in \tableofcontents in case e.g. section without description.
In case a section, subsection etc just has a tag and not a description in the \tableofcontents there is nothing mentioned. With this patch the tag is used as description when no description is present, as is done on other places as well (consistency and no empty items in the \tableofcontents).