Last comment of \code{.f90} missing
In case e.g. an number of Fortran variables are documented as an example in a \code block and the last line looks like:
INTEGER :: var !< documentation
The "!< documentation" is missing as the last block was not rendered.
Note: STRIP_CODE_COMMENTS has to be NO as otherwise the code comment will anyway, correctly,  be absent.