Support dotfile files with same names from different directories
In case \dotfile is used with the name name but from a different directory the result is only 1 output file. This is corrected with this patch.
In the html version the warning in respect to systems without svg lead to problems (<p> tag inside <p> tag), this has been made consistent between \dotfile and \dot ...\enddot