LaTeX problem with 2 consecutive single quotes
In LaTeX 2 consecutive single quotes are joined to 1 double quote. This leads in texts where 2 single quotes are intended to a strange view. An example is the initial value for variables. E.g. in Fortran: character(len=10) :: A = ''  (i.e an empty string which can either be defined as '' or as "") or in C: char *sq  = "''". In the source listings this is OK but in the description text it was not OK.