Fortran color code of END
The color for the single END in Fortran code was of the color of the flow type entities though for all the flow entities the entity name is mandatory. For the entity statements of some keywords e.g. SUBROUTINE and FUNCTION the entity name is not mandatory with the END statement. The color of the single END statement has been changed from the flow type to the normal keyword type.