Bug 722940 - classes with same name in different namespace adding to same group
If two classes with the same name, but in different namespaces (so different qualified names) were added to the same group, only one of them would end up in the group.
The problem is caused by qualifiedName of the class not containing the qualified name but the class name when a class is added to the group.
A workaround is to use the name attribute of the class, which does contain the qualified name.