Fortran: warning message about not documented module member
When having a not documented MODULE member it is mentioned that a member in a 'namespace' is not documented. For Fortran this is a bit strange as Fortran does not know the term namespace, modules are handled as namespaces (in the pre 1.8.9 version a message was given about 'class', which is equally confusing).
This patch changes in this case 'namespace' to 'module'.