Enable relative paths referenced source files
This is a regression on pull request 210: Create an easy possibility to take a snippet from the current file.
In case the source file, as referenced through 'this' in the snippet command, is in the current directory or in a subdirectory pull request worked fine, but as soon as the source file is in a directory referenced through ../... the parser returns an absolute path name instead of a relative path and file is not found, resulting in a warning plus empty snippet part.
We now check whether or not is an absolute file name in which case it is checked if the file is found or not and further handled.