Support generating index-color PNG files
Based on the request http://stackoverflow.com/questions/27036474/how-do-make-doxygen-generate-index-color-png-files for index colored png files, this functionality will be included in the newer versions (2.39) of dot.
The functionality -Tpng:cairo:gd will automatically be used by doxygen, though the file extension would be wrong (.png:cairo:gd).
With this patch the :cairo:gd is stripped from the image extension, but not from the -T of the dot command.
In the doxywizard the different possibilities are supplied as well. Note however that the DOT_IMAGE_FORMAT name is not checked when read from the Doxyfile so it is possible to use other renderer / formats as well due to the generic solution with getDot ImageExtension.