Better error message in case of IDL inconsistency
In case of an IDL internal doxygen inconsistency ("error: Internal inconsistency: namespace in IDL not module, library or
constant group ", see forum report: http://doxygen.10944.n7.nabble.com/Internal-inconsistency-Which-file-td7057.html) it is hard to find out which file causes the problem.
In this case the comment block with @namespace was the "problem").

This patch solves this by adding the file and line number of the source file causing the problem.