Crash in case of non generic interface in Fortran
In case of a non generic interface in Fortran it happened that  doxygen crashed due to the fact that the pointer for setting the endBodyLine does not exist. The line number is only defined for GENERIC interfaces and therefore it should not be set anyhow when ending a  non generic interface.