Bug 746614 - Bogus warning regarding nested comments
In case the previous file gives a warning regarding the wrong used nested comments and the next file does not have any doxygen comment there is still the message that there is a nested comment but even without probable line numbers.
The variable counting the nested comment level was not reset, this has been corrected with this patch.