Problem running tests under Windows
The redirection used in the runtests.py script was based on *nix type systems. Corrected for windows.
Windows has a different line ending than *nix, resulting in a problem with diff unless the options -b -w are used.
Some tests had as directory for some paths a directory . this has been corrected to $INPUTDIR