Copy user EXTRA_FILES at the end to overrule files generated by default by doxygen
In case the user wants to overrule by means of a HTML_EXTRA_FILIES or LATEX_EXTRA_FILES a default file from doxygen (e.g. a logo or arrowdown) this is not directly possible as the files are again overwritten by doxygen.
Placing the copy commands at the end and in this list the generateTreeViewImages before the the user files a user can overrule default doxygen files