Stop when doxygen configuration file ("Doxyfile") cannot be read
When the configuration file cannot be read a default configuration is used ad as the error message "error: cannot open file `Doxyfile' for reading" is show at the top of the output this is easily overseen.
By stopping the doxygen process it is made clearer that there is an error.