Bug 753909 - Copy and paste of code fragment from CHM merges all pasted text into single line
Complete code fragment was pasted in one line (chm, HTML OK), with the change in the ccs file this problem is overcome. Result was that between multiple code lines on the 2nd and following line a extra space appeared at the beginning of the line, this is overcome by placing all relevant div statements on one line.

(based a.o. on http://stackoverflow.com/questions/19099873/how-can-i-use-css-to-insert-a-line-break-after-but-not-before-an-element)