Title in case of USE_MDFILE_AS_MAINPAGE
In case USE_MDFILE_AS_MAINPAGE is used the title in the HTML output is the project name followed by the word documentation and in the index of LATeX / rtf the value is "Main Page", this is even the case when a level 1 header is given. This is a bit contrary to the handling non main pages.
With this patch in case of USE_MDFILE_AS_MAINPAGE and a level 1 header in the beginning of the main page file this level 1 header is used title