Fortran module private variables and EXTRACT_PRIVATE = NO
From the message in the doxygen forum(http://doxygen.10944.n7.nabble.com/Fortan-problems-td7434.html and http://doxygen.10944.n7.nabble.com/Fortran-module-private-variables-and-EXTRACT-PRIVATE-NO-td7423.html)
B. If EXTRACT_PRIVATE = NO, module private functions are not listed (as expected) but module private variables are (bug!).
C. Implicit PRIVATE/PUBLIC/PROTECTED statements in a module work as expected but are not "appended" to function/variable documentation. This is only a nuisance.

Problem had a common cause, the protection was hard set to 'Public' instead of the read value (analogous to static)