Markdown with @cond and @endcond
In case markdown files are processed and they contain @cond and @endcond these are replaced wit C-style end and begin comments (and code in between is correctly removed). The C-Style comments should not be placed as they are not understood in a markdown context (analogous to what happens in Fortram, Pythone etc.).
Note: in principle @cond should not be used as a markdown file is one "comment" block and thus @if should be used, though in any case no unwanted characters should be added.