Correct / set types for python variables
The bool python type was not handled.
In case of the initialization by means of a method call a variables was was aid to be a tuple (e.g. env= os.environ.copy())
Methods with initialization with defaults were shown without arguments in the brief description (e.g. def create_dir(dir, create=True):)
Arguments with values in method calls were seen as variables variables in case of continuation lines (e.g.
     parser.add_argument("--https_proxy", nargs="?", help="Proxy to be used for https requests",
                        action="store", const="myproxy.domain.com:8080")
  in this case action was seen as variable).