bunch of performance improvements
In one test (with an older version from Debian unstable, I
didn't re-profile with the unmodified master branch),
Config_getBool("FULL_PATH_NAMES") accounted for 25% of the
runtime according to Valgrind, mostly because of
FileDef::name(). While the real number is probably smaller
(I didn't turn on cache simulation), I think it makes sense
to cache the value.

FilePairDict::compareValues() is pretty hot, so avoid computing
orderLo there if it's not needed.

reNumberNode is really hot, and if QT was built with bounds
checks enabled, every call to .at() costs a bounds check.
Apparently gcc can't figure out how to get rid of it, at least
not at -O2 without LTO. Therefore, use memset() to get rid of
the repeated bounds check and make writing the zeroes faster.

FileDef::name() has a 6% cost after the previous improvements
because every invocation of it unnecessarily clones the string.