performance: store node numbering for graphs in maps
When processing a large source tree, for every callgraph, a
gigantic QArray is allocated to store the number mappings.
For just the fraction of the Linux kernel that's required
for a specific Android device, that's around 260MB RAM that
are allocated, zeroed and deallocated for every single
callgraph.

Just use a map instead.