performance: run dot faster
Use the -O feature of doxygen that's available since 2006:

https://github.com/ellson/graphviz/commit/574e183b6caed4954702929a952c06c59a141a9e

This of course breaks compatibility with super-old systems.
Is that a problem? If so, a new flag could be added, I guess?
But in that case, the flags should IMO default to the newer
behavior IMO. Or support could probably be autodetected.