Fortran inline source code and crash on Linux
This patch fixes a problem on Linux where in a number of cases doxygen crashed when an unnamed interface existed and the routine itself was in the same source file.
Furthermore the end line of routines is set correctly so  inline source code in Fortran will be possible as well.