Latex page numbering
In the documentation generated by doxygen the table of contents does not start with page number i but with iii, by setting the pagenumbering}{roman} after the clearemptydoublepage titlepage this is corrected. In the beginning of the documentation we set the page numbering to "alph" so it won't interfere with the numbers in the table of contents and the real documentation.

For the doxygen manual similar reasoning is true, furthermore the table of contents was on the left side page (when printed double sided), yhis is corrected as well.