Fix STRIP_FROM_PATH when running from drive root
When running doxygen from the drive root, the STRIP_FROM_PATH,
variable could get an extra slash. As a result no paths were
stripped. This fix makes sure that a duplicate slash is not added
both for the automatic generation, and when you specify a relative
path that maps to the drive root in the configuration.

Also change some tabs to spaces