Fix SHOW_NAMESPACES=NO to hide all namespaces docs
This change will exclude the generated namespace documentation from
the resulting output files. Currently, even if this option is enabled
the namespace documentation is included.
The generation of namespace indices is inhibited which also leads to
an error when generating a LaTeX documentation (`namespaces.tex' not found),
except when the indices are hidden (`LATEX_HIDE_INDICES=YES`).

This also fixes [Bug 728649](https://bugzilla.gnome.org/show_bug.cgi?id=728649).