make compound.xsd deterministic again
Some validating XML parsers (e.g. the Qt5 one) complain about
ambiguity and duplicate elements in the compound.xsd.

For example:

    Error XSDError in file:xml/compound.xsd, at line 781, column 20:
      Complex type docCaptionType has non-deterministic content.
    Exception: XSD xml/compound.xsd is invalid

And after this is fixed:

    Error XSDError in file:xml/compound.xsd, at line 683, column 20:
      Complex type docParaType has duplicated element image in
      its content model.
    Exception: XSD xml/compound.xsd is invalid

The first is due to those elements being generated from a table where the
`bareName` is not unique. Resulting in a choice element of `docTitleCmdGroup`
containing the element `trademark` twice. This change filters
out the duplicates.

The second is due to the choice element of `docCmdGroup` referencing
`docTitleCmdGroup` - which defines element `image` - while also locally
defining the `image` element. This change remove the redundant local
definition.