Bug 766059 - DOT_PATH not expanded
The original problem in this bug request was that forward slashes were used in the PATH instead of backslashes.
In the command shell this does not result in problems, but when calling a program from within doxygen this leads to the problem that the executable cannot be found.
In this patch the PATH variable is, just for the process, changed in such a way that ther are back slashes instead of forward slashes (Windows only, not on Cygwin).