Fix plantuml generation issue
Commit 62f09103b77b9e5b4093d248563e2f9d76b65728 which
merged pull request 538 introduced a bug in the generation
of plantuml diagrams for other plantuml sources than the
htmldocvisitor. In the previous commit the baseName argument
was wrongly interpreted as file name without path. This
commit fix this issue so the behaviour should be the same
as before, while still also having the support for
PlantUML diagrams in Qt compressed help files.