Fix bug #774432: Autogrouped only one level of entities
Preivously, compound entities autogrouped using \{ and \} will have all
their descendents entities included in the group as well. This makes the
generated module page very noisy.

This patch maintains a stack of QCString, corresponding to the name of
the autogroup for the compound entities currently being recursively
parsed. Then upon `initGroupInfo`, an autogroup is added to the entry's
list of groups only if the autogroup is different from its parent
compound entity's autogroup.