More correct logic for checking parent autogroup
Adjusted code according to Doxygen's parsing logic. Previously it was
buggy.

Also, namespaces are ignored when deciding if an entry's parent is
already included in the autogroup. This is a more correct behavior since
namespaces are open and can be extended at other places. When a
namespace is included in an autogroup via `\{` and `\}`, it really
the content of the namespace defined within those braces that should be
included in the namespace, and not the namespace itself, which can
contain other entites defined elsewhere.

For example, consider the following snippets

```cpp
/// \defgroup autogroup
/// \{
/// desc of x
namespace x {
    /// desc of y
    namespace y {
        /// desc of a
        struct a {
            /// desc of a::b
            struct b{};

            /// desc of a::c
            int c;
        };
    }
}
/// \}
```

Then the module page for `autogroup` will include `x`, `x::y`,
`x::y::a`, but not `x::y::a::b` or `x::y::a::c`;