Improve brazilian portuguese translation
I reviewed all the brazilian portuguese translation, and have found some flaws. Most are semantical errors, and there are a few syntactical errors, too. Also, I changed a few translations to better ones, for the sake of comprehension.

There exists a rule of thumb in portuguese that foreigns words have to be italicized, so I put a <em> tag on 'Singleton' word, as it hasn't a good translation to portuguese.