Call endMemberItem consistently.
When generating man pages, enums erroneously output their description
on the same line as the definition, with a trailing '"'. This is due to
endMemberItem being called after the member description is emitted, rather
than before, as it is elsewhere.

This patch moves the call to the same place where it is called elsewhere.

Bad output:
enum foo { FOO = 0 } Description. "

Good output:
enum foo { FOO = 0 }
	Description.


Signed-off-by: Alastair D'Silva <alastair@d-silva.org>