Adding the following configuration options:
SHOW_USED_FILES_TXT
If set to NO the text in front of the used files list is omitted

SEPARATE_MEMBER_PAGES_TITLE
If set to NO the title area of a member description is omitted.

For separate member pages (i.e. SEPARATE_MEMBER_PAGES = YES) I have added a header section
in the same way it is available for class pages. This looks more consistent.
Furthermore the class and member function name is displayed in that header and is a link back to the class page it belongs to.