Expose underlying enum type in the XML output.
The <type> element (used for function return type and variable type) is
now reused also for underlying enum type.

C++03-style enum types that don't have the underlying type specified
have the <type> tag as well, but empty to indicate that the underlying
type is unspecified.

Added a new test case for this.